Fixed TrackingRequest doing page view requests
Page view request were broken because it tried to submit "page" as the
request type when it did a page view. For some reason, when you submit
"page" the page view never gets counted. It disapears into the neather.
However, you don't even need to include it at all because if you don't
specify a type then the documentation says it will be a page view.

- Removed a ton of duplicated code using three helper functions
- Fixed TrackingRequest as a page view
- Changed properties to functions (they should be!)
- Removed unused code