Removed app.config file dependency
Now the google analytics ID is passed in to the Factory. Optionally, if
you do use the app.config file you can just use
ConfigurationSettings.GoogleAccountCode to get it from your app config
without writing any new code. The only difference is that it does not do
this automatically now.

The reason for this is that I wanted to use this library in a plugin where
I do not wish to have app.config file dependencies.